Initial commit:
The bot is functional ( sometimes ) and is able to control Discord through a specified list of commands ( 2 at this time ).

Problems:
1:Output from ChatGPT does not always contain a JSON format or its output contains unwanted text
Fix: Modify the prompt.txt to not require a JSON as it seems too hard for the AI to
complete. And instead request a list of commands in a String with each command separated
by a new line, and their parameters by spaces. ( literally a breakthrough for me LOL )

2: When the AI tries to communicate, its responses are too short
Fix 1: See if the previous fix would make it better, and if not then... ¯\_(ツ)_/¯